broke test into components
